Created a front-end sign-up page
Created a front-end Sign-up page using React, which captures user's details and sends a POST request to our server.
Implemented a loading spinner which activates when the server is processing the sign-up request.
Implemented a message system to show success or error messages to the user. Also added styling to everything in the sign up component.