Changed Modal Functionality
Removed Email Prop: I removed the email prop from the ConfirmationModal component, as it is no longer necessary for the email confirmation process.

Confirmation Code Handling: Updated the handleEmailConfirmation function. The function no longer requires the email in the body of the POST request. Now, it only sends the confirmationCode to the server, aligning with the updated backend logic.

Error and Message Handling: Included state handling for errors and messages to ensure user feedback is properly delivered. It sets error messages if the server response is not OK or if there's any exception while processing the request. Conversely, it sets success messages if the response is OK.

Loading State: Incorporated a loading state to provide visual feedback to the user during the email confirmation process. It displays a loading spinner while the confirmation request is in progress.

Redirection: After successful email confirmation, the modal closes and navigates the user to the home page. The duration for closing the modal has been adjusted from 2 seconds to 5 seconds, giving users sufficient time to read the success message.

State Updates in Input Field: Included state updates in the input field to keep track of the confirmationCode input by the user.