Implement and style Menu database for both backend and frontend
I've successfully created, implemented, and styled the Menu database for the site. On the backendm I designed the Menu database schema to hold info for all menu items. This includes, names, descriptions, prices, image URLs, quantities, and categories. I also set yp routes for fetching this data, which includes categorizing the items based on their category for easy access on the frontend. On the frontend side of things i made state for menu items and fetched this data from the backend using 'fetch'. The data is receieved and stored in this state. Then the menu items are rendered based on their category. I also styled the menu page and i'm pretty happy with the results so far. I also made a script called seeder.js in the backend directory that I used to fill menu database with the menu items. I could use this same script later to add more items if need be easily.