Finished styling the cart and made a basic checkout component
In the header, there's a shopping cart button that shows the total number of items currently in the cart. When clicked, a modal opens to display the cart's content. Each item in the cart is shown in its individual container with its image on the left. To the right of the image, the item name is displayed and below it, the item's price is shown. On the far right of each cart item container, there are controls to increase or decrease the quantity of the item.

The cart provides a subtotal, which is the total cost of all the items. Below the subtotal, depending on the number of items in the cart, there's either a 'Continue to Checkout' button or a 'Go to Menu' button. If there's at least one item in the cart, the 'Continue to Checkout' button is displayed. If the cart is empty, the 'Go to Menu' button is shown, allowing the user to return to the shopping menu.

On the checkout page, there's a form where users can fill out their information to complete the purchase. It includes fields for the user's name, address, city, state, and zip code. After filling in these details, the user can click the 'Place Order' button to complete the purchase. The checkout page is accessible only if there's at least one item in the cart.