The application's authentication process has been expanded with a JWT (JSON Web Token) validation feature. The client-side is now capable of validating user tokens automatically at the start of each session. This is achieved by introducing a new validateToken function in UserContext.jsx. This function checks for the presence of a token in local storage and sends a POST request to the server to validate it. The login and validateToken functions in loginControllers on the server-side have also been updated to accommodate this change. User interfaces for signup and login are now wrapped within a UserProvider context in App.jsx to ensure user data is shared across components.
Server-side changes:

A new validateToken function has been added to loginControllers.js. It attempts to verify the token received in the Authorization header of the request. If the token is valid and corresponds to a user, the user's data is returned in the response. If the token is invalid or the user does not exist, an error is returned.
The validateToken function is exported from loginControllers.js and a new route (/login/validate-token) is added to handle POST requests with this function.
Client-side changes:

UserContext.jsx:
A new validateToken function is created that fetches the token from local storage, sends it to the server for validation, and sets the user data if the token is valid.
The validateToken function is called when the application starts using the useEffect hook.
App.jsx:
The UserProvider context has been added to the application, wrapping all components to ensure user data is shared across components. The login modal is now conditionally rendered based on the isLoginModalOpen state.
The setIsLoginModalOpen function is passed as a prop to the Header and HomePage components, allowing them to open the login modal.
Component directory changes: The ConfirmEmail, SignUp, and LoginModal components have been moved to a new directory named User-Components.