Added conditional rendering of elements based on whether a user is logged in or not.
Changed the positioning of things in the header so that the cart button is now located in the nav bar. When a user is logged in the sign up and sign in buttons in the nav bar aren't rendered and a Welcome message with the user's name and logout button are rendered instead. In the Hero section on the home page, if a user is logged in the sign up/ sign in buttons aren't rendered and a welcome message is rendered there instead with the user's name.