Created a UserProfile component and add routes on the back-end related to the user's profile
The user profile URL is a unique 9-digit code that is generated when a user creates an account. This code is passed as a parameter and used in the url for the profile. I also changed the back-end related routes to start with /api in order to distinguish them from front-end routes.