Fleshed out the UserProfile page and created an edit profile modal.
Fleshed out the user's profile page with user information such as name, email, and password. I also created a new Edit Profile modal that will appear when the user clicks the edit icon on the profile page. I Added sections for the user's address and billing information that will be fleshed out with the user's actual information when I get around to implementing that. I also moved the logout button from the header into the the user's profile page. I also changed the JWT to expire to 12 hours from 1 hour. Fixed an issue that occurred when a user logs in but their information isn't displayed correctly until the page is manually reloaded.