Added password visibility toggle button to all password fields
Implemented a password visibility toggle button for all passwords fields in my application. The user clicks the eye icon and the password fields will be visible until they click it again. I also added labels for the input fields in the LoginModal component to better distinguish them from one another.  I still have 3 more things left to implement regarding signing in and singing up. First: change the error messages so that they are more specific to what actually went wrong. Second: Add a resend reset / confirmation code option so that the user can resend these emails containing the codes if they didn't get the first one for some reason. Third: When a user signs up and makes an account they should automatically be logged in with that account instead of having to go back to the home page and logging in manually.