Added clear error messages and input validation to LoginModal.jsx
Made the error message when something goes wrong while logging in clear. Also added input validation to the input fields in the login modal so that the user cannot submit empty forms and is told what they need to submit in order to log in. I also brought in the loading spinner that i've used in other components to the login modal. Also added a login successful message that will be displayed before the user is sent back to the home page after logging in.