Implemented Front-End Functionality for Resending Confirmation Emails
In this update, I've developed the front-end interfaces and implemented fetch requests for resending confirmation emails in the ConfirmEmail and PasswordChange components.

For the ConfirmEmail component, a new interface element was added to allow users to trigger a request to resend the confirmation email. Upon clicking this element, the application sends a fetch request to the server endpoint responsible for resending confirmation emails. The server processes the request and sends a new confirmation email to the user's registered email address.

Similarly, for the PasswordChange component, I've added a mechanism to allow users to request a password reset email if their reset code is expired or misplaced. This mechanism is designed to initiate a fetch request to the server endpoint that handles password reset email resending when triggered.