Implemented email confirmation check during login and added ConfirmEmail component
In this commit, I've introduced changes to the back-end /login route. Now, it verifies if a user's email is confirmed when they attempt to login. If a user's email is not confirmed, the server responds with a status code of 403 to indicate that access is forbidden until the email is confirmed.

To handle this new response on the front-end, I updated the LoginModal component. It now catches a 403 status code response and communicates to the user that they cannot login until their email is confirmed. To aid in this process, a "Re-send confirmation email" link is shown to the user.

Upon clicking this link, the ConfirmEmail component is rendered where the user can enter the confirmation code they received. To accommodate this new functionality, I had to adjust the handleResendEmail function. It's now split into two parts: one part in the ConfirmEmail component that handles the user interaction, and another standalone function handleResendEmail that sends the request to the server to re-send the email.

To enable this functionality in multiple components (LoginModal and SignUpModal), I moved handleResendEmail to the parent App.jsx file and passed it down as a prop.