Changed the route and request to include the user's JWT
I changed the route controller function, updateUser, to include a check on the user's JWT. I made this change to ensure that only the user can make the request to change their name. I also changed the request on the front-end to include the user's JWT in the Authorization header of the request.