Updated loginController and added shippingInfoController
In this commit, I've updated the loginController to now check if a user is a guest or not upon login attempts. This allows for an added layer of validation, preventing guest accounts from signing in. If the user attempting to log in is a guest, the controller will now return a 401 error and an explanatory message.

I've also introduced a new file, shippingInfoController.js. This new controller manages user shipping information. It offers the ability to fetch a user's shipping information, set a default shipping address, add a new address, and delete an existing one. Each of these operations is secured with JWT based authorization, ensuring that users can only manipulate their own shipping information. The logic to automatically make the first address added by a user their default one, if none exists, has also been implemented. In addition, when a default address is deleted, the controller automatically assigns the default status to another address if available.