Moved UserProfile component, added new components ShippingInfoModal and CreateAddressModal, and rearranged buttons in UserProfile.
The UserProfile component has been moved from the User-Components folder to the Page-Components folder.

A new component, ShippingInfoModal, has been added. This modal is accessible from the UserProfile and it displays the addresses associated with the user's account. It also includes an "Add Address" button for adding new addresses.

I've introduced another new component, CreateAddressModal, that's used for adding new addresses. This modal is brought up when the "Add Address" button in ShippingInfoModal is clicked. It includes a form for inputting address details, performing client-side validation, and submitting the new address to the server. The new address, if successfully added, is then appended to the user's addresses in the context. In addition, there's a disclaimer and a checkbox within the form that the user must acknowledge before they can submit a new address. The disclaimer informs the user that they shouldn't enter any real payment or address information as this site is just a project that will be in my portfolio and it does not process real transactions.

Within the UserProfile component, I've moved the EditProfile button to be grouped with the Shipping Info and Billing Info buttons.