Updated UserContext.jsx to handle and provide user shipping addresses.
UserContext is now used to share not only the user's data, but also the user's addresses across components.

The useState hook is now used to initialize a state variable addresses along with the setAddresses function to update it.

The login function comment now clarifies that it stores the user's data and token in state and local storage.

A logout function has been implemented which, alongside removing user data from state and clearing the token from local storage, also removes the addresses from state.

The original useEffect hook that was used to validate the JWT when the application starts, now also fetches addresses for the user, if the JWT is valid. If the address response is ok, the addresses are extracted and the addresses state is updated.

The UserContext.Provider now also provides addresses and setAddresses to its children. This means the children components can now access and update the addresses as needed.

This is part of an ongoing effort to implement the shipping information feature in the application. The provided shipping info will be accessible by components that need it, like ShippingInfoModal in the UserProfile, and in future, the Checkout component once it's fully implemented.