Added stateAbbrev and countryAbbrev to Prisma schema and adjusted form fields.
I updated the Prisma schema model for ShippingInfo to include two additional fields: stateAbbrev and countryAbbrev. These fields hold the abbreviations for states and countries respectively. The stateAbbrev field is nullable, which means it's only populated when the selected country is the United States.

I have also revised the form fields for state and country, transforming them from input fields to select menus. The country field now makes use of a file, COUNTRIES.js, in the helpers folder which contains names and abbreviations for all countries.

Regarding the state field, I made it conditionally rendered. It now checks whether the selected country is the United States. If it is (which is the default state), the state field changes to a select menu populated with all US states, drawn from the US-States.js file in the helpers folder. If the United States is not selected as the country, the state field changes back to a regular input field, and the label alters to 'State / Province / Region'.

I also adjusted the POST request on both front-end and back-end to accommodate these changes.