Created new Utils folder in the back-end and rateLimit.js file
Moved the seed.js to the utils folder and created a new rateLimit.js. This rateLimit.js takes the rateLimit function that was used in the userRoutes.js for limiting the amount of requests a user can make and puts in it's own file within the new utils folder so that i can be easily re-used by importing it in any other routes i may want to limit the amount of requests later on down the line. This keeps my code DRY!