Added front-end PATCH request and message state for updating default user address
This commit includes the addition of a function `setDefault` in the component which sends a PATCH request to the back-end, setting a user's address as default. This function is invoked when a user chooses to make an address their default one.

The front-end now includes a `message` state, which is used to communicate back to the user about the status of their request.

Also, the request made by this function is rate limited at the back-end to prevent the user from making too many requests in a short time span.