Added address deletion function and automatic message clearing in the ShippingInfoModal component.
Address Deletion Function: A new function called deleteAddress has been added to the ShippingInfoModal component. It checks if the addressId is valid, confirms with the user if they are deleting their last address, and then makes a DELETE request to the server. If successful, it removes the deleted address from the local state and sets a success message. If not, it shows an error message.

Automatic Message Clearing: A new useEffect hook has been added which automatically clears the message state after 15 seconds. This is helpful to avoid any lingering messages from previous actions. For now, this change is just specific to the ShippingInfoModal component, I still need to consider whether or not to use it in other components in the application that also have message states as not all messages (like those set during signup or login) should be subject to a timeout.