Improved user experience by reflecting updated default address immediately after deletion without requiring page refresh.
Backend Changes (ShippingInfoController):

Despite having a system in place to automatically set a new default address when the existing one was deleted, the updated default address information was not being communicated to the front-end, causing a discrepancy between the backend and frontend states. To fix this, I modified the server response upon successful address deletion.

Now, when a default address is deleted, the server responds with a message indicating successful deletion, and if a new default address was set (which would be the case if there are remaining addresses), it includes the information of the new default address in the response. This allows the front-end to immediately reflect the new default address.

Frontend Changes (ShippingInfoModal):

With the backend now sending the new default address in the response, I needed to update the front-end to handle this new information.

Upon receiving the response from the server, I updated the addresses in the context by removing the deleted address and checking if the response included a new default address. If it did, I iterated through the remaining addresses and set the new default address to true for the address with the matching ID. This ensures that the new default address is immediately displayed as the default on the UI without requiring a page refresh.