Changes have been made to the Prisma schema and new front-end components that handle billing and card information
Starting with the Prisma schema, I introduced comments to each model, clearly defining how each one relates to others. These comments are meant to provide clarity on how the models are interconnected so that it's easier to understand the data structure.

Another addition to the schema is the new 'cardInfo' model. This model has been designed to store fictitious card data for demonstration purposes. In real-world application development, storing actual card data this way wouldn't be secure or recommended due to payment data security standards (PCI DSS). However, given the context of this application, the model serves as a placeholder to simulate the card creation process.

The 'billingInfo' model was adjusted to incorporate this new 'cardInfo' model. This change allows the consolidation of billing and card information under the same data model. It's a way to tie the information together.

On the front-end two new components, BillingInfoModal and CreateCardModal, were created. These components provide similar functionality to the existing 'ShippingInfoModal' and 'CreateAddressModal' but cater to billing and card information. They were created to have the same look and feel of these earlier components.

THINGS LEFT TO DO:
The next step is to create functions that generate bogus card information for the user. Along with that, I need to add controller functions and routes for both the 'billingInfo' and 'cardInfo' models on the back-end. Once those are complete I will then need to create the corresponding front-end requests for each of the new routes.