Created controller functions for user's cardInfo CRUD operations, defined their routes, and implemented an encryption/decryption helper.
In the recent changes, I created controller functions handling all CRUD (Create, Read, Update, Delete) operations for a user's cardInfo. These functions are pretty similar to the controller functions used for handling a user's shipping information.

I also implemented an encryption and decryption helper for card information. This ensures that all sensitive card information is encrypted before being stored and is decrypted only when it is required to be accessed.

While the current use of this helper is for card information, I could also use it for encrypting user passwords. Right now the passwords are already being encrypted with their own process but I could use this new helper instead in order to keep my code DRY!

However, the immediate next step will be to focus on creating front-end requests to these newly established routes for handling card information.