Added card information to UserContext, implemented a POST request in CreateCardModal, displayed payment cards in BillingInfoModal, and reorganized front-end components and stylesheets.
Integration of Card Information in UserContext: I introduced the cardInfo state variable in the UserContext. This is populated through a GET request that fetches and stores the information for use across child components. This ensures that data can be conveniently accessed in multiple parts of the application like the BillingInfoModal and the upcoming Checkout component.

POST Request in CreateCardModal Component: I incorporated a POST request into the CreateCardModal component. It takes the bogus card details that are generated for the user and creates cardInfo for the user in the database.

Display of Payment Cards in BillingInfoModal: In the BillingInfoModal, I added functionality to display the user's stored payment cards. This feature mirrors the display of addresses in the ShippingModalComponent, providing a consistent user experience.

Front-End Directory Restructuring: Lastly, I've reorganized the structure of components and CSS stylesheets. I've created two new directories - Modal-Components and Modal-Styles, where I've placed all modal components and their styles. This organization makes it easier to manage the increasing number of modal components in the application.

Future plans include adding requests for setting a default card and deleting cards from the user database. Once these features are implemented, the focus will shift towards developing the checkout process. The end of this project is now on the horizon!