Implemented a staged checkout process in the application.
Made the Checkout component stage-based, separating it into "Shopping Cart", "Shipping", "Billing", and "Order Review" steps.

Introduced a new state, currentStage, to manage which stage to display.

Shifted cart-related code to a new component, ShoppingCart, which is the default stage on checkout page load.

Added an ItemRecommend component to appear with the shopping cart.

Established a new directory, Checkout-Components, to house all checkout related components, and a corresponding styles directory.