Refactoring of App.jsx to move hooks and data fetching operations into separate files for better organization, access, and reusability across different components.
In this update, the logic for managing cart states, resending confirmation emails, and fetching menu items was removed from the App.jsx component. This was done to address code clutter and accessibility concerns.

A new folder, 'hooks', was created. Each major functionality was assigned to its own hook:
useCart: Handles the cart state and includes functions such as addToCart, removeFromCart, and clearCart.
useEmailResend: Provides the functionality to resend confirmation emails.
useMenuItems: Fetches the menu items to be displayed in the menu and used in the ItemRecommend component.
These hooks were imported back into App.jsx, and their functionalities were utilized within the component. The reorganization and compartmentalization of the code improve maintainability and readability.

The menu items were initially fetched in the Menu.jsx component, causing accessibility issues when these items were needed in the ItemRecommend component. By moving the fetching logic to the App.jsx component using the useMenuItems hook, we ensured that the menu items could be fetched at a higher level and passed down as props to both the Checkout (which then passes them to ItemRecommend) and the Menu components. This improvement allows for a more streamlined data flow and efficient use of resources.