Simplified Checkout into two main stages
Checkout.jsx is the main component for the checkout process, now divided into two stages. The first stage shows the user's shopping cart, and the second uses the CheckoutDetails component to collect user info for checkout and order placement. A breadcrumbs function is included to help users track their stage and navigate back if needed.