Introduced Shipping component for address selection and management.
The Shipping component serves as the interface for users to select or manage their shipping addresses during the checkout process. It utilizes the UserContext to fetch existing addresses and allows for the addition of new ones through a modal. The component also features an expandAddresses state to toggle between a summary and detailed view of addresses. A useEffect is employed to automatically set a default address if one exists. Users can change their address selection through radio buttons, and new addresses can be added via the "Add new address" button, which opens a CreateAddressModal. As it stands right now this component only works when a user is logged in, and has addresses saved in context. I will need to return to this later and fix it. So, it's not done yet!