Added Billing component for payment method selection and management.
The Billing component is designed to handle the user's payment method during the checkout process. It fetches existing card information from the UserContext and allows users to add new cards through a modal. The component features an expandCards state to switch between a summary and detailed view of payment methods. A useEffect sets a default card if one exists. Users can select their payment method through radio buttons, and new cards can be added via the "Add new card" button, which opens the CreateCardModal which is the same one I used in the BillingInfoModal. As it stands right now, just like the Shipping.jxs component, this component only works if a user is both logged in, and has some card information already stored for them. I will need to return to this later and fix it!