Introduced ShippingMethods component for shipping option selection.
The ShippingMethods component is part of the CheckoutDetails and is responsible for allowing users to select a shipping method. It offers multiple shipping options like "Standard," "Express," and "Overnight," along with a "Free" option for orders over $100. The component uses a useEffect to set a default shipping method based on the user's eligibility for free shipping and the total cost of the cart. Users can change their shipping method through radio buttons, and the shipping cost is dynamically calculated and updated using the calculateShippingPrice function. Messages are displayed to inform users about free shipping eligibility. I'm pretty sure this component will work without requiring a user to be logged in but I will need to do some more testing on it when I return to fix it's siblings the Billing and Shipping components later.