Implemented OrderSummary component for displaying order costs and details.
The OrderSummary component serves as the final stage in the checkout process, providing a view of the order's financial details. It currently displays the subtotal, tax, shipping costs, and the total order amount. The component accepts various props like totalCost, totalItems, shippingCost, shippingMethod, cardSelection, addressSelection, and user. These props are crucial for future development, as they will be used to send a request to place the order. The component uses a useEffect to calculate and update the tax amount and the total order cost dynamically. It also includes a button that will be used to actually place the order. Future updates will integrate front and back-end routes to finalize the order placement.